+++ include/linux/mempool.h
@@ -97,10 +97,10 @@ void mempool_kfree(void *element, void *pool_data);
 
 #define mempool_init_kmalloc_pool(_pool, _min_nr, _size)		\
 	mempool_init(_pool, (_min_nr), mempool_kmalloc, mempool_kfree,	\
-		     (void *)(unsigned long)(_size))
+		     (void *)(uintptr_t)(_size))
 #define mempool_create_kmalloc_pool(_min_nr, _size)			\
 	mempool_create((_min_nr), mempool_kmalloc, mempool_kfree,	\
-		       (void *)(unsigned long)(_size))
+		       (void *)(uintptr_t)(_size))
 
 /*
  * A mempool_alloc_t and mempool_free_t for a simple page allocator that
